requirements.txt
pymsteams[async]
azure-identity
azure-keyvault-secrets